requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
mysql-connector-python==8.1.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
